+++ test/unit/test-md-ref.js
@@ -12,7 +12,7 @@ test('Reference with a single ref', function (t) {
   const input = `
     See reference[^1]
 
-    [1]: some text`
+    [^1]: some text`
 
   const output = inlineHtml`
     <p>See reference<a href="#reference1"><sup>1</sup></a></p>
@@ -20,7 +20,11 @@ test('Reference with a single ref', function (t) {
       <p><sup id="reference1">1</sup>some text</p>
     </section>`
 
-  t.equal(parseToHtml(input), output, 'Output is valid')
+  const opt = {
+    allowFootnote: true,
+  }
+
+  t.equal(parseToHtml(input, opt), output, 'Output is valid')
   t.end()
 })
 
@@ -28,8 +32,8 @@ test('Reference with 2 refs', function (t) {
   const input = `
     See this[^1] and that[^2]
 
-    [1]: ref one
-    [2]: ref two`
+    [^1]: ref one
+    [^2]: ref two`
 
   const output = inlineHtml`
     <p>See this<a href="#reference1"><sup>1</sup></a> and that<a href="#reference2"><sup>2</sup></a></p>
@@ -38,15 +42,19 @@ test('Reference with 2 refs', function (t) {
       <p><sup id="reference2">2</sup>ref two</p>
     </section>`
 
-  t.equal(parseToHtml(input), output, 'Output is valid')
+  const opt = {
+    allowFootnote: true,
+  }
+
+  t.equal(parseToHtml(input, opt), output, 'Output is valid')
   t.end()
 })
 
 test('Reference with 2 refs not in order', function (t) {
   const input = `
     See this[^1] and that[^2]
-    [2]: ref two
-    [1]: ref one`
+    [^2]: ref two
+    [^1]: ref one`
 
   const output = inlineHtml`
     <p>See this<a href="#reference1"><sup>1</sup></a> and that<a href="#reference2"><sup>2</sup></a></p>
@@ -55,7 +63,11 @@ test('Reference with 2 refs not in order', function (t) {
       <p><sup id="reference2">2</sup>ref two</p>
     </section>`
 
-  t.equal(parseToHtml(input), output, 'Output is valid')
+  const opt = {
+    allowFootnote: true,
+  }
+
+  t.equal(parseToHtml(input, opt), output, 'Output is valid')
   t.end()
 })
 
@@ -63,8 +75,8 @@ test('Reference with text ref', function (t) {
   const input = `
     See this[^one] and that[^two]
 
-    [one]: ref one
-    [two]: ref two`
+    [^one]: ref one
+    [^two]: ref two`
 
   const output = inlineHtml`
     <p>See this<a href="#reference1"><sup>1</sup></a> and that<a href="#reference2"><sup>2</sup></a></p>
@@ -73,7 +85,11 @@ test('Reference with text ref', function (t) {
       <p><sup id="reference2">2</sup>ref two</p>
     </section>`
 
-  t.equal(parseToHtml(input), output, 'Output is valid')
+  const opt = {
+    allowFootnote: true,
+  }
+
+  t.equal(parseToHtml(input, opt), output, 'Output is valid')
   t.end()
 })
 
@@ -81,8 +97,8 @@ test('Reference with missing ref key', function (t) {
   const input = `
     See this[^one] and that
 
-    [one]: ref one
-    [two]: ref two`
+    [^one]: ref one
+    [^two]: ref two`
 
   const output = inlineHtml`
     <p>See this<a href="#reference1"><sup>1</sup></a> and that</p>
@@ -90,7 +106,11 @@ test('Reference with missing ref key', function (t) {
       <p><sup id="reference1">1</sup>ref one</p>
     </section>`
 
-  t.equal(parseToHtml(input), output, 'Output is valid')
+  const opt = {
+    allowFootnote: true,
+  }
+
+  t.equal(parseToHtml(input, opt), output, 'Output is valid')
   t.end()
 })
 
@@ -98,7 +118,7 @@ test('Reference with missing ref text', function (t) {
   const input = `
     See this[^one] and that[^two]
 
-    [one]: ref one`
+    [^one]: ref one`
 
   const output = inlineHtml`
     <p>See this<a href="#reference1"><sup>1</sup></a> and that<a href="#reference2"><sup>2</sup></a></p>
@@ -106,7 +126,11 @@ test('Reference with missing ref text', function (t) {
       <p><sup id="reference1">1</sup>ref one</p>
     </section>`
 
-  t.equal(parseToHtml(input), output, 'Output is valid')
+  const opt = {
+    allowFootnote: true,
+  }
+
+  t.equal(parseToHtml(input, opt), output, 'Output is valid')
   t.end()
 })
 
@@ -114,7 +138,7 @@ test('Reference with complex texts', function (t) {
   const input = `
     See reference[^1]
 
-    [1]: some *italic* and **bold** and a [link](address)`
+    [^1]: some *italic* and **bold** and a [link](address)`
 
   const output = inlineHtml`
     <p>See reference<a href="#reference1"><sup>1</sup></a></p>
@@ -122,7 +146,11 @@ test('Reference with complex texts', function (t) {
       <p><sup id="reference1">1</sup>some <em>italic</em> and <strong>bold</strong> and a <a href="address">link</a></p>
     </section>`
 
-  t.equal(parseToHtml(input), output, 'Output is valid')
+  const opt = {
+    allowFootnote: true,
+  }
+
+  t.equal(parseToHtml(input, opt), output, 'Output is valid')
   t.end()
 })
 
@@ -130,10 +158,10 @@ test('Reference with callback', function (t) {
   const input = `
     See reference[^1]
 
-    [1]: some text`
+    [^1]: some text`
 
   const opt = {
-    allowReference: true,
+    allowFootnote: true,
     onReference: node => {
       t.notEqual(node, null, 'Parameter is populated')
       t.equal(node.tagName, 'A', 'Tagname is valid')
@@ -149,11 +177,11 @@ test('Reference with allowReference to false', function (t) {
   const input = `
     See reference[^1]
 
-    [1]: some text`
+    [^1]: some text`
 
   const output = inlineHtml`
     <p>See reference[<sup>1</sup>]</p>
-    <p>[1]: some text</p>`
+    <p>[<sup>1</sup>]: some text</p>`
 
   const opt = {
     allowReference: false,